fix(h2-events): bound the per-stream telemetry map so proxied h2 traffic can't pin RAM
H2EventLog records frame-level telemetry for proxied (untrusted) HTTP/2 traffic.
The event ring was capped at kMaxEvents, but the per-stream summary map (m_streams)
had NO cap and NO eviction -- every stream a connection ever opened stayed
resident, closed streams were never erased, and a status/RST/close for an unseen
stream auto-created a permanent entry. So a long session (or a server cycling
stream ids / spraying RST) grew the map without bound, contradicting the header's
own "attacker bytes can't pin RAM" promise (which the comment also described with
a per-conn ring / kMaxEventsPerConn that never existed).

Add a kMaxStreams cap with closed-first, then oldest-openedAtMs eviction, applied
in every note*() path that can insert. Extracted as pure H2Logic::capStreamSummaries
+ clampTelemetryField so the policy is unit-tested (new Tests/h2_logic, 20
assertions: closed-first, oldest-among-open, orphan-openedAt-0 shed first,
converge-to-cap, defensive max<=0 no-op, field clamp incl. a 1 MB :path).

Also from the same review:
- framesIn/framesOut were int -> signed-overflow UB on a very-long-lived stream;
  widened to qint64 to match the byte counters.
- :method/:path were stored verbatim (a malicious upstream could send a multi-MB
  :path); now length-clamped to kMaxTelemetryFieldLen.
- noteRequestHeader/noteResponseStatus/noteStreamClosed never emitted
  eventsChanged(), so the UI didn't refresh on a clean close/status; now they emit
  (outside the lock, matching noteFrame/noteRstStream).
- Corrected the false header comment.

ctest 61->62; probe_smoke 156/156; h2_events.cpp recompiled clean.

+++ Src/BackEnd/Proxy/h2_events.hpp
@@ -13,9 +13,12 @@
 //   * Surface SETTINGS / WINDOW_UPDATE asymmetry that flags weird WAF
 //     behaviour
 //
-// Per-connection state is held in H2EventLog as a ring of
-// kMaxEventsPerConn events; rolling buffer behaviour mirrors
-// WsRepeater so attacker bytes can't pin RAM.
+// State is held in H2EventLog as a single process-wide ring of the most
+// recent kMaxEvents frame events (rolling buffer, mirrors WsRepeater) PLUS a
+// per-stream summary map. BOTH are bounded so proxied (untrusted) traffic
+// can't pin RAM: the event ring drops its oldest, and the summary map is
+// capped at H2Logic::kMaxStreams with oldest/closed-first eviction. Stored
+// :method / :path strings are length-clamped (H2Logic::kMaxTelemetryFieldLen).
 
 #include <QByteArray>
 #include <QMap>
@@ -46,8 +49,8 @@ struct H2StreamSummary {
     int     status     = 0;
     qint64  bytesIn    = 0;
     qint64  bytesOut   = 0;
-    int     framesIn   = 0;
-    int     framesOut  = 0;
+    qint64  framesIn   = 0;   // 64-bit: a long-lived stream can exceed INT_MAX frames
+    qint64  framesOut  = 0;
     quint32 lastError  = 0;
     qint64  openedAtMs = 0;
     bool    closed     = false;
